evas: small 25% speed up of textgrid (we are still on average 10ms slower than urxvt).
This patch reuse the same Evas_Text_Props across the textgrid (Later we could do that
across the canvas itself, but no test case yet, so not yet). It delay the real destruction
of Evas_Text_Props after the rendering cycle, so they will really die only if they are
not used during the latest frame. We could delay it a little bit more, but that does
the job at the moment.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@72510 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33