elev8: Implement __require__()
This should be used sparingly, because it loads a module in the current
context. If modules are badly written, they can override symbols in the
current content and all hell can (or rather will) break loose.

To create a JavaScript module correctly, wrap its contents around an
anonymous function and call afterwards, like this:

   (function() {
        var private_var = 42;
        global_var = 2.718281828;
   })();

This way, in your main module, you'll have access to ``global_var'' but
``private_var'' won't be accessible.

The dangerous-looking name is by design. This function is scary and should
not be used unless you really know what you're doing. You've been warned.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@72618 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33