eet: Go back to plain header check for zlib as band-aid
Older version of zlib, used in Fedora 10 and Ubuntu 10.04, does not have a pc file. It would be
better to bring this into m4 macro but I fear touching it. Vincent, it would be great if you could
handle this as well.

Without this you get this error:
configure: error: "Cannot find zlib.PC. Make sure your PKG_CONFIG_PATH
environment variable contains include lines for the location of this
file"

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@73138 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33