From: Doyoun Kang <doyoun.kang@samsung.com> Subject: [E-devel] [Patch][e] add code to check e_config->geometry_auto_move before moving a window in e_border.c
In e, there is a configuration value - geometry_auto_move - which
determines whether e moves the window to useful geometry automatically or not.
In my opinion, if this is 0, then e doesn't move the window
automatically. However, e always moves the window automatically
because there is no code to check the geometry_auto_move value in
_e_border_eval(). So, I added code to check
e_config->geometry_auto_move in _e_border_eval().



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@73191 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33