From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] diskselector, duplicate icon for rounded item
minor fix to diskselector to repeat icons right.

The diskselector has round mode which can be set by
elm_diskselector_round_enabled_set().
In case of round mode is enabled, the icon does not display properly
because diskselector replace the icon for the additional items - not copy
and paste. To enable round mode, diskselector adds over/under itmes(additional
itmes) by using normal mode items. So there is a icon to set two places in
diskselector. But diskselector does not duplicate the icon.. this makes item
without icon but there should be.. I'm not sure the best way to duplicate icon.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@73397 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33