From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] popup - focus chain, accessibility
you would be in rush because of freezing.. but there is still hot. -_-;
i'd like to raise an issue about focus chain on the popup. The focus would
go around only in the popup. (IMHO)
but current popup gives a chance to get the focus to other widget which is
not in the popup.
please check the attachment which has accessibility feature also and give
your opinion. thanks.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@74467 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33