Elm: Fix ticket #1245. Mouse cursor does not change before window resizing.
Previously, we only changed the mouse cursor when the border frame was
actually clicked. Now we can change it when the mouse moves over the
border/frame areas (like in X11). This adds a callback from the edc to
know when and where the mouse moves (with respect to the frame border
only). (IE: When the user moves the mouse over the bottom portion of
the border, the edc will let us know and we can change pointer
accordingly).




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@74925 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33