eina: add EINA_SENTINEL to help use of variadic functions
Certain types of variadic functions use NULL as the last argument instead of a
string format (printf-like). Functions like these are: execl and execlp.

We are in feature freeze, but I believe this is small and simple enough to slip
in with no headaches. These functions are being used in the new edbus library
and it would be good to have it supported in eina now.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@75271 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33