Ecore_Evas (wayland_shm): Fullscreen mode bug fixes
Only take min & max properties into account if the window is not fullscreen. 

Be sure to call update_size before we attach the buffer, as the buffer attach 
code makes use of the allocated size to damage the surface.

Add trapping for fullscreen window in the configure callback. This
(along with coming commits) allows us to actually do fullscreen now :)

During window configure, when we check for maximized, we should check
to be sure it is not fullscreen also before adjusting window size.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@75293 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33