From: Jihyeon Seol <jihyeon.seol@samsung.com> Subject: [E-devel] [Patch] elc_multibuttonentry resize patch.
Currently, there is a problem in the logic of getting the size in the
MBE's
box.

For each line, the height of the line is calculated according to the
biggest
item's height 
and is applied to the box size, but problem occurs when the newly added
item's height is larger than the one before.

For example, when the entry is located in the front of the last line, 
the line's height is determined by the entry. 

When the button is added by the user's input at this state, 
because button's height is larger than the entry's height 
I have modified each line to have the equal height.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@75453 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33