[elm] Rework focus_region and on_focus_region  hooks.
The former is now issued for all scrollables, with the behavior it had
before, and the latter is a new virtual base function, with a default
implementation leading to values matching the old case where one
didn't implement the hook.

Please help me test if everything is OK -- it seems so, here.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@75904 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33