Massive interface improvement, seriously massive.
Ok, this is much better.  I'd like to thank elm_grid
for making all this possible.  Without you being pretty
much the only widget in all of Elementary that makes any
sense to me, none of this would be possible.

Anyway, give it a try, let me know if it's good, bad,
terrible, etc.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@76327 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33