From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] calendar - add access feature, access - add _elm_access_edje_object_part_object_unregister();
the attachment has accessibility feature which is for the
elm_calendar. and
moreover..
it would be better to keep one more api for the access which name is
_elm_access_edje_object_part_object_unregister();
in the case of calendar item, its text part could be set with empty
value
in run time(dynamically), even though it had a value (1~31) previously.
so if there is an empty field(item), then previously registered item
should
be unregistered. the api would be useful not only this case but also
others.
then, please review the patch and give feedback. thanks.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@76502 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33