From: Jihyeon Seol <jihyeon.seol@samsung.com> Subject: [E-devel] [Patch] multibuttonentry - data corruption issue of the box list
here is a problem in the data corruption of the box list in
_box_layout_cb() of the multibuttonentry.

EINA_LIST_FOREACH is used to move and resize the objects. 
but if box unpack/pack is called in the EINA_LIST_FOREACH loop, 
the data of the box list will be corrupted.

This problem can not be solved immediately due to the structural
problems of
the MBE. 
So the attached patch, a workaround, avoids calling unpack/pack in the
EINA_LIST_FOREACH loop.

I am in the process of refactoring MBE but in the mean time, 
this patch will resolve the above issue.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77192 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33