eeze/sensor: Fix big memleak that happened on every eeze use.
Really free the list with sensor modules when shutting down. No idea how I
dropped that as I found it in one old local version I had. Thanks to Mike
for the notice and valgrind for the tooling.

Also make sure we don't corrupt the list before we can free it.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77291 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33