From: Kim Shinwoo <kimcinoo.efl@gmail.com> Subject: [E-devel] [patch][elementary] diskselector - bug fix
diskselector has two bugs.

1. on round enable mode
elm_diskselector_item_append(); does not call elm_box_pack_*() for
over_items and under_items which are added when the display item is
bigger
than 3.
because over_items and under_items is not added, the diskselector shows
just 4 items + half item(NOT 5 items), in case of display item number
is 5
- elm_diskselector_display_item_num(obj, 5); -.
this bug could be found only if u use as following sequence.
   a. elm_diskselector_round_enabled_set(obj, EINA_TRUE);
      b. elm_diskselector_item_append();
      
      2. on the round disable mode + bigger than 3 display item
      even though display item is bigger than 3 on the round disable
mode, only
two blank items are added - left_blank, right_blank.
because of this, the first item could not be selected. means.. the first
item cannot be located in the middle of diskselector.
if u scroll the diskselector to the left-most, then the second item is
selected as below.

   [blank] [ 1st ] [ 2nd ] [ 3rd ] [ 4th ]
   
   so if the display item is bigger than 3, more blank items would be
   necessary.
   include lines to resolve above 1st bug, the patch would resolve
this issue
also.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77740 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33