efl: cleanup to remove stuff that is not in single-tree yet.
Remove every mention to libraries that are still not in the
single-tree build such as emotion, efreet and ethumb.

Keep it simple and just add them back when they are merged.

Added similar variables for all of evil, eina, eo and eet. Keep them
in the same order for now.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77767 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33