efl: introduce build profile, simplify options.
Introduced --with-profile={dev,release} that will simplify how to set
build options of EFL.

NOTE-1: specific e17 benchmark is now gone, it will always be built
and is the default benchmark for eina. If we want to have a faster
benchmark in the future, just add a command line option for
eina_suite.

NOTE-2: valgrind build is broken as it needs -fPIC. Will get to it
later. Likely someone needs to revisit the eina mempools for valgrind
and other basic tools (eo? likely evas).



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77771 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33