efl: simplify crypto options.
Introduce a new --with-crypto={openssl,gnutls,none} that will allow
one to choose the EFL cryptographic system.

If set to gnutls or openssl, cipher and signature will be
enabled. Otherwise it's disabled.

NOTE: gnutls is trying to cope with old API and that sucks. Should we
just drop the support for old gnutls and bump the required version?



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77789 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33