PROTO/elocation: Turn meta provider creation back on.
Create a geoclue meta provider on startup which can be used instead of
specific providers like UbuntuGeoIP, gpsd or such. Geoclue internally
decides and aggregates the best data for this providers given your
requirements.

I'm still unclear how to destroy this object at the end as their is no
function for it and the reference counting does not help to free it
automatically.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77894 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33