EasyUI: Profile view, to display user's profile
The most important thing here is the approach used to display data
on the screen. Now, it's possible to return (on listController itemAtIndex)
an item with `style` property set to `'full'`. In this case, GenController
will display the `content` property of the item. This property must contain
a widget, like `Table`. The twitter app is a test case, then. Maybe this
change should be separated from twitter, but, hey, how to see if it works?


Patch by: ederson <ederson@profusion.mobi>



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77913 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33