EasyUI: Improvements on ListController 'LoadingItem'
When showing or hiding the loading item, some 'scroll,*' events are
fired. A timeout was added when showing the loadingItem (there was a
timeout before, but just when hiding the loadingItem).
The timeout is also longer than before, but it's just an attempt to
avoid problems when the animation takes longer.


Patch by: ederson <ederson@profusion.mobi>



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77916 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33