From: rustyBSD <rustyBSD@gmx.fr> Subject: [E-devel] [ecore] Recursive rm
I looked at the ecore_file_recursive_rm() function
(in ecore/src/lib/ecore_file/), and I wonder why this
is so ugly/complicated.

We are doing a readlink() and two stat(). Why not simply
do a lstat() ?

It takes less memory and it's simplier. Here is a patch.

(And readlink() is ssize_t)



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@77979 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33