edbus: Fix type check of solo structs
When reading messages that contain structs the type that comes from the
wire (and obtained with dbus_message_iter_get_arg_type()) is 'r'.
However 'r' never appears in a signature and dbus_signature_iter_*
returns '('.

Ref: http://dbus.freedesktop.org/doc/dbus-specification.html#message-protocol-signatures

"STRUCT has a type code, ASCII character 'r', but this type code does not
appear in signatures. Instead, ASCII characters '(' and ')' are used to
mark the beginning and end of the struct. So for example, a struct
containing two integers would have this signature:"

Patch by: José Roberto de Souza  <zehortigoza@profusion.mobi>



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@78072 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33