evas: give a Warp 10 engine to XPM loader !
NOTE: If anyone know the moron who did think it was a
great idea to make XPM the Freedesktop recommanded 
choice for icons, please give him a huge SPANK for me !

NOTE: There is in fact no reason why CServe2 and default
loader code are not shared. The number of different line
is small. I highly advise we make the effort to merge
them back soon.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@78424 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33