E17:
I. Simplification, andif we cannot stat(), the best
thing to do is toswitch to a copy-delete operation.

II.Just stuff...

III. There is a problem with _e_fm_op_random_char().
When wewant to randomize a string we do srand(time())
foreachcharacter, so if the file is small enough
to be deleted inless than a second, it's not randomized.
And even if it's bigger, it's not goodly randomized.Sorry.

Patch by Maxime Villard rustyBSD


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@78469 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33