ephysics: support disabling stacking based on Z axis
Since now we support movement on Z axis the evas objects
associated to bodies need to be re-stacked to be below / above
each other.

But if Z axis movement is disabled, no cloth is used, ...,
it can be safely disabled, since evas objects won't overlap.
And it will save performance.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@78491 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33