EasyUI: Add Tablet Mode
Summary:
In this mode, some things should behave differently. For instance, the split
panel mode will always be expanded, and the button to show it won't be
generated.

To make an application use tablet mode, set the EASYUI_UI_MODE variable to
'tablet'.  Setting it to anything different will use the default, 'phone'
mode.  This is an environment variable since the application launcher will
set that to reflect the environment the application is running on.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@78777 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33