edbus: fix checking if root objects are our son
strncmp'ing is not sufficient because it whould fail in case an object
like "/org/bla" is registered (and in conn->root_objs) and we were
registering "/org/blatest/". We would incorrectly put "/org/bla" as son
of "/org/blatest". Fix this by looking up the next char, checking if
it's a slash.

Also re-format the code a little bit.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@79179 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33