edbus: check signature earlier and fail on invalid signature
The problem is: if there's any arg with invalid signatures, we cannot
continue adding the interface. Just fail with a log message so user will
notice it.

This means we need to check if signals are valid much earlier. Actually
before even adding the object/interface.

Also change the iteration on signals so we don't use a needless index.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@79184 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33