e: use subdir-objects only on modules
autopoint/gettext stuff isn't happy at all with subdir-objects option in
configure.ac. We need this option to compile the modules or otherwise we
have a conflict in the generated objects.

big spankies to the gettext crowd.

"fix" it by only using subdir-objects in src/modules/Makefile.am, where
it's really needed.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@79208 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33