efl: Fix eet test suite.
Cedric suggested that we migth not be able to open the cert files. And he
was right. The CERT_DIR define checked for an absolute path in TESTS_SRC_DIR
but that get set to .. in Makefile_Eet.am as it sits in src/. Now the define
evaluated to src/tests/eet/.. where no cert files could be found.

I could not see a good reason for the suffix here so I just removed it.
Vincent, if there was a reason I was not able to see please fix it in a
different way. It was the least invasisve fix I could see.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@79838 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33