[elm] Introducing a new widget, along with its infrastructure -- prefs.
The prefs widgets aims to aid with the implementation of
preference/configuration windows/UI elements in Elementary-based
applications (think of Enlightenment configuration dialogs,
elementary_config, etc).

Prefs is a widget that populates its view with widgets
bound to data types (following the instructions of a ".epb" file that
describes a set of items) and handles the storage/restoration of such
data on a configuration file automatically.

There's also the prefs_data handle, which is the one dealing with
user saved data for a given epb defaults set.

The documentation on the new widget is rich (we have examples and even
an EPC reference) and there's a new test entry for it.

I'm blogging about it soon, with screeshots and more details.

Enjoy.

ps.: This is a team work by Murilo Belluzzo, Ricardo de Almeida and me.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@79909 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33