elm now supports edbus v2, deprecated v1 function.
v1 is now deprecated (EINA_DEPRECATED) but still there, should still
work and not break any existing app.

v2 is now there as well, all software is being ported to use it
now. Just Enlightenment itself will still ship with v1 and as soon as
we release it will go v2, we have the patches here.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80110 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33