directfb says bye...
After agreement in the mail list, core developers agree to remove this
engine that was not being supported for a long time.

Given that most operations Evas uses are not accelerated in DirectFB,
or at least hardware that exclusively supports DirectFB, it's better
for those people to just use Evas/Ecore software (buffer) rendering
and expose DirectFB's framebuffer as destination surface.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80232 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33